app/requirements.txt
fastapi
uvicorn
sqlalchemy
sqlmodel
passlib
PyJWT